backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sympy==1.12
numpy==1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
Pillow==10.3.0
httpx==0.27.0
groq==0.9.0

# Auth packages
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
email-validator==2.1.1